requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
google-generativeai>=0.7.0
faiss-cpu>=1.8.0
numpy>=1.26.0
pypdf>=4.2.0
python-multipart>=0.0.9
httpx>=0.27.0
pytest>=8.2.0
pytest-asyncio>=0.23.0

# Optional local embedding fallback (no Gemini key required)
# sentence-transformers>=2.7.0
